The scores work !!

+++ Cards_Game/Model/CardMatchingGame.h
@@ -33,7 +33,7 @@
 
 
 
-@property (nonatomic, readonly) NSInteger score;
+@property (nonatomic, readonly) int score;
 @property (nonatomic) int maxOfMatchingItems;
 @property (nonatomic) int numberOfCardsLeftToMatch;
